moved UI code to new MaltNodeUITools file
